Move the mujoco independent part of the gjk algorithm into a separate file
